requirements.txt
pyyaml==6.0.3
platformdirs==4.5.0
typer==0.20.0
pydantic==2.12.4
pandas==2.14.3
pyarrow==22.0.0
influxdb-client==1.35.0